Move forms and tabs from metadata.py to tabs.py
The forms were moved to tabs as that's where they're now being
used, also to avoid circular imports.

Each tab is now a separate widget with buttons pointing to
the forms in that tab, they all conform to the same interface.

This makes the tabs modular and also easier to find and fix
issues when they arise.

I've only added the forms that exist and commented out the ones
yet to exist.
